Refactor (APIMF-3491): extracted json pointer property parser
